some small changes.
